requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==3.0.3
frozenlist==1.4.1
future==1.0.0
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
line-bot-sdk==3.11.0
MarkupSafe==2.1.5
MouseInfo==0.1.3
multidict==6.0.5
packaging==24.1
pillow==10.3.0
PyAutoGUI==0.9.54
pydantic==2.8.0
pydantic_core==2.20.0
pyfiglet==1.0.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyocr==0.8.5
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
pytweening==1.2.0
requests==2.31.0
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.2
wrapt==1.16.0
yarl==1.9.4
